refactor: start map refactor
