+++ src/styles/mapbox-gl.scss
@@ -13,9 +13,9 @@
   width: 100%;
   height: 100%;
 }
-.mapboxgl-canary {
-  background-color: salmon;
-}
+// .mapboxgl-canary {
+//   background-color: salmon;
+// }
 .mapboxgl-canvas-container.mapboxgl-interactive,
 .mapboxgl-ctrl-group button.mapboxgl-ctrl-compass {
   cursor: -webkit-grab;
@@ -707,3 +707,5 @@ a.mapboxgl-ctrl-logo.mapboxgl-compact {
     display: none;
   }
 }
+
+.mapboxgl-ctrl { display: none !important; }
