+++ src/styles/style.scss
@@ -121,4 +121,5 @@ img { line-height: 0; }
 @import './countries-list.scss';
 @import './tabs-group.scss';
 @import './mapbox-gl.scss';
+@import './map-custom-styles.scss';
 @import '../keyboard/keyboard.scss';
